SERVER-3060 partial fix
This commit fixes the specific case described in SERVER-3060, but
this is only one of many similar issues in the code.  The problem
is that an exception in any of the low-level conversion routines in
scripting\engine_spidermonkey.cpp when called by scripting\sm_db.cpp
will cause a jump out of the SpiderMonkey script interpretor unless
the exception is caught inside the native code that SpiderMonkey has
called.  This commit handles bad parameters passed to mongo_find(),
but the same problem exists with mongo_update, mongo_insert, and
mongo_remove.  There may be other cases as well.

In this case, a numeric zero is passed to Convertor::toObject()
(the code is f = c.toObject( argv[2] );) and toObject throws an
exception, tossing us all the way out to shell/dbshell.cpp, leaving
the SpiderMonkey's internal state in a bad condition.  Catching the
exception in mongo_find() and returning JS_FALSE lets SpiderMonkey
keep its state in working order.