SERVER-2942 Hook functions in the Windows Import Address Table
In the Windows version of mongod and test, hook the RtlCreateHeap and
NtAllocateVirtualMemory functions in ntdll.dll through the Import
Address Tables of kernel32.dll and/or kernelbase.dll.  Use RemapLock
to block memory allocations during the period when remapPrivateView()
has the private view unmapped so that other threads can't grab any of
the address space required for the private view.