SERVER-6082 skip cleanup step on Windows, it generates exception
The ShardingTest.prototype.stop function is not up to the task of
cleaning up the state of things established by sharding/remove2.js,
and it tries to delete files that are in use by running servers.
See SERVER-6094 for some information on the failure modes.  The
error is seen while server shutdown is already in progress on Linux
and the original (zero) return code is returned so the test is
considered to have succeeded.  On Windows, the error is seen by
JavaScript when it tries to delete a file that is in use, which
aborts the test.  This change just skips the "stop" call on Windows.