SERVER-7497 refactor and rationalize mongo/util/file.h; create file.cpp
Remove the FileInterface base class.  Move the code for non-trivial members
of the File class into a new mongo/util/file.cpp.  Make the behavior of
Windows and non-Windows versions of functions match.  Test the return status
of every OS API and report all errors in detail.  Make file.h not depend on
previously included headers for symbols it uses.