SERVER-7678 Reduce document count in mrShardedOutput.js for 32-bit
Reduce the number of documents inserted per iteration from 100,000
to 30,000 in jstests/sharding/mrShardedOutput.js when run on a 32-bit
system.  Do not print the list of inserted documents when the count
is wrong, just display the documents that are missing.  Add logging
so we can see where we are in the test if something goes wrong.