SERVER-5099 Use Unicode command line in mongod.exe, mongos.exe and test.exe
Use the WindowsCommandLine class to convert Windows Unicode (UTF-16) text
on the command line to UTF-8 for internal use in mongod.exe, mongos.exe
and test.exe.  Implement support for the third parameter of main()/wmain()
and convert the Unicode enviroment strings to UTF-8 as well.  Use the
environment parameter in mongo.exe, mongod.exe, mongos.exe and test.exe.