SERVER-8741 Add new sharding test
Add test based on jstests/sharding/mrShardedOutput.js to help debug
missing documents on sharded inserts.  This test is similar to
mrShardedOutput.js, but has no MapReduce step.  Includes new
library function to analyze splits and migrates relative to the
insertion of a document for cases when documents go missing.