SERVER-9786 Link to strcasestr at runtime for Solaris
For the Solaris/SmartOS build, do not make direct call to strcasestr,
which is present in Solaris 11 but not in Solaris 10.  Instead, see if
if is available in libc.so.1 at runtime and either call it or call an
emulation.  Use the emulated version on Windows, which is also missing
this function.