CS-7316 Add partial support for backtrace in Solaris 10
Dynamically link to backtrace(), backtrace_symbols() and backtrace_symbols_fd()
at runtime on Solaris.  Emulate the backtrace() function if it is not found at
runtime.  This allows stack addresses to be displayed.  The emulation routines
for backtrace_symbols() and backtrace_symbols_fd() are stubs, so symbolic
information will not be shown in this version.